% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-functions.R
\name{osc}
\alias{osc}
\alias{hillside}
\alias{hill}
\alias{flat}
\alias{steady}
\alias{same}
\alias{exp2}
\alias{exp10}
\alias{recip}
\alias{magnitude}
\title{definitions of core modeling functions}
\usage{
osc(x)

hillside(t)

hill(x)

flat(x)

steady(x)

same(x)

exp2(x)

exp10(x)

recip(x)

magnitude(x)
}
\arguments{
\item{x}{Generic argument to function}
}
\description{
definitions of core modeling functions
}
