% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensions.R
\docType{data}
\name{dimensions}
\alias{dimensions}
\alias{mydimensions}
\alias{mydim}
\alias{available_dimensions}
\title{Utilities for typesetting dimensions}
\format{
An object of class \code{character} of length 20.
}
\usage{
mydimensions

mydim(what)

available_dimensions()
}
\arguments{
\item{what}{Typesets a dimension using exponents on the base dimensions.}
}
\description{
Utilities for typesetting dimensions
}
\keyword{datasets}
