% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRho.R
\name{getRho}
\alias{getRho}
\title{Extract estimates of the intra-class correlation from a glmmTMB
model object with beta-binomial error.}
\usage{
getRho(obj)
}
\arguments{
\item{obj}{\pkg{glmmTMB} model object with betabinomial error}
}
\value{
Vector with as many elements as there are observations,
holding the estimates of \eqn{\rho}.
}
\description{
The intra-class correlation is calculated as
  \eqn{(1+exp(\theta))^{-1}}, where \eqn{\theta} is the
  estimate given by the formula specified in the argument
  \code{dispformula}.
}
\examples{
{
codling1988 <- qra::codling1988
codling1988$gp <- with(codling1988, paste0(Cultivar,rep))
ge16xl.TMB <- glmmTMB::glmmTMB(formula=cbind(dead,total-dead)~0+Cultivar/dose+(1|gp),
dispformula=~Cultivar+poly(dose,2), family=glmmTMB::betabinomial(link='logit'),
data=subset(codling1988,dose>=16))
rho <- getRho(ge16xl.TMB)
}
}
