% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphResid.R
\name{graphResid}
\alias{graphResid}
\title{Draw plots f residuals from model fitted to insect mortality data}
\usage{
graphResid(df, subSet = NULL, logScale = FALSE, dosevar = "logCT",
  Res = "Res", Rep = "Rep", span = 0.85, byFacet = ~Species,
  layout = NULL, maint = "MeBr", pch = 16, ptSize = 2,
  xzeroOffsetFrac = 0.08, xlab = expression(bold("CT ") * "(gm.h." * m^{    
  -3 } * ")"))
}
\arguments{
\item{df}{Data frame from which data will be taken}

\item{subSet}{NULL, or an expression, such  as for example
\code{expression(LifeStage=='Eggs')}) that evaluates to a logical that
specifies the required data subset.  If not NULL then the subsetting
information is pasted on after the main title}

\item{logScale}{Logical, indicating whether the dose ($x$-variable)
is on a log scale.}

\item{dosevar}{Character; column holding "dose" values}

\item{Res}{Character; name of column holding residuals}

\item{Rep}{Character; NULL, or column holding replicate number, within panel}

\item{span}{Span for fitted smooth(s).  To omit smooth(s), set to zero}

\item{byFacet}{Graphics formula specifying factor combination that
determines panel}

\item{layout}{Graphics formula that can be supplied to \code{grid_facet}}

\item{maint}{Main title}

\item{pch}{Plot shape(s), by default 16}

\item{ptSize}{Pointsize, by default 2}

\item{xzeroOffsetFrac}{$x$-axis zero offset fraction, required when
scale is logarithmic, by default 0.08}

\item{xlab}{Expression specifying x-axis label}
}
\description{
Datasets that are in mind hold, for each replicate of
each combination of each of a several factors (e.g.,
species, lifestages, temperatures), model residuals for
each of a number of values of "dose".
}
