% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_index.R
\name{heat_index}
\alias{heat_index}
\title{Calculate the heat index based on temperature and humidity}
\usage{
heat_index(temperature, humidity)
}
\arguments{
\item{humidity}{relative humidity in percent, but as a decimal, e.g. 0.50}

\item{temerature}{in deg C}
}
\value{
Heat index in degrees C
}
\description{
Implements a polynomial approximation to the Steadman table, per Rothfusz.
Out of bounds inputs still generate a result, although it is not meaningful
owing to the nature of polynomials explosive growth
}
