\name{kerrich}
\alias{kerrich}
\docType{data}
\title{
Kerrich Coin Toss Trial Outcomes
}
\description{
A data set containing 2,000 trials of coin flips from statistician John Edmund Kerrich's 1940s experiments while imprisoned by the Nazis during World War Two.
}
\usage{data("kerrich")}
\format{
  The format is:
List of 1
 $ : chr [1:2000] "0" "0" "0" "1" ...
}
\source{
\url{https://en.wikipedia.org/wiki/John_Edmund_Kerrich}
}
\references{
Kerrich, J. E. (1950). An experimental introduction to the theory of probability. Belgisk Import Company.
}
\examples{
data(kerrich)
}
\keyword{datasets}
