% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-modeling-profiles.R
\name{osc_graph}
\alias{osc_graph}
\title{All of these are hidden, to be called with triple colons, e.g.  QRA:::osc_graph()}
\usage{
osc_graph()
}
\description{
All of these are hidden, to be called with triple colons, e.g.  QRA:::osc_graph()
}
