% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra.r
\name{qra}
\alias{qra}
\title{Quantile Regression Average
Calculates a quantile regression average for forecasts.}
\usage{
qra(
  forecast,
  target,
  group = c(),
  model = "Quantile Regression Average",
  per_quantile_weights = FALSE,
  enforce_normalisation = TRUE,
  intercept = FALSE,
  noncross = TRUE,
  ...
)
}
\arguments{
\item{forecast}{a data.table representing forecast; this is expected to
have been created using \code{\link[scoringutils:as_forecast]{scoringutils::as_forecast()}}}

\item{target}{the target for which to create the quantile regression
average. This should be given as a vector of form \code{column = target},
where target is the value of column that represents the target. Note that
the column named here cannot be a grouping variable.}

\item{group}{any columns wihch to group a vector of character vectors (e.g.,
"horizon", "geography_scale", etc.) indicating columns in the
\code{forecasts} and \code{data} data frames; by default, will not group
anything, i.e. create one ensemble model}

\item{model}{the name of the model to return; default: "Quantile Regression
Average"}

\item{per_quantile_weights}{logical; whether to estimate weights per
quantile}

\item{enforce_normalisation}{logical; whether to enforce quantiles}

\item{intercept}{logical; whether to estimate and intercept}

\item{noncross}{logical; whether ot enforce non-crosssing of quantiles}

\item{...}{passed to \code{\link[quantgen:predict.quantile_ensemble]{quantgen::predict.quantile_ensemble()}}; of particular
interest might be setting \code{iso = TRUE} for isotonic regression}
}
\value{
a data.table representing the forecasts forecast, but with
\code{model} set to the value of the `model parameter. This will be in the
forecast format produced by \code{\link[scoringutils:as_forecast]{scoringutils::as_forecast()}}
}
\description{
Quantile Regression Average
Calculates a quantile regression average for forecasts.
}
