% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra.r
\name{qra_create_ensemble}
\alias{qra_create_ensemble}
\title{Helper function to create a QRA ensemble}
\usage{
qra_create_ensemble(preds, qra_res, iso = FALSE)
}
\arguments{
\item{preds}{predictions}

\item{qra_res}{QRA result}
}
\value{
tibble
}
\description{
Helper function to create a QRA ensemble
}
