% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra.r
\name{qra_estimate_weights}
\alias{qra_estimate_weights}
\title{Estimate weights for QRA for a data frame containing \code{model}, \code{centile},
\code{boundary},  \code{value}, \code{interval} columns}
\usage{
qra_estimate_weights(x, per_centile_weights, enforce_normalisation)
}
\arguments{
\item{x}{input data frame}

\item{per_centile_weights}{if TRUE, separate weights are calculated for
different centiles}

\item{enforce_normalisation}{if TRUE, normalisation (weights sum to 1) is enforced}
}
\value{
data frame with weights
}
\description{
Estimate weights for QRA for a data frame containing \code{model}, \code{centile},
\code{boundary},  \code{value}, \code{interval} columns
}
\keyword{internal}
