% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra.r
\name{split_forecast}
\alias{split_forecast}
\title{Split forecast}
\usage{
split_forecast(forecast, forecast_unit, target)
}
\arguments{
\item{forecast}{a data.table representing forecast; this is expected to
have been created using \code{\link[scoringutils:as_forecast_quantile]{scoringutils::as_forecast_quantile()}}}

\item{target}{the target for which to create the quantile regression
average. This should be given as a vector of form \code{column = target},
where target is the value of column that represents the target. Note that
the column named here cannot be a grouping variable.}
}
\value{
a list with forecast (the forecast with models removed that have
missing forecasts, and the target forecasts removed)
}
\description{
Splits forecast into data and forecast, excludes any models aren't complete,
and splits out the forecast target
}
\keyword{internal}
