% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra.r
\name{to_matrix}
\alias{to_matrix}
\title{Helper function to convert a data frame of forecasts into a list of matrices}
\usage{
to_matrix(x)
}
\arguments{
\item{x}{input data frame}
}
\value{
list of matrices
}
\description{
Helper function to convert a data frame of forecasts into a list of matrices
}
\author{
Sebastian Funk
}
\keyword{internal}
