% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractLT.R
\name{extractLT}
\alias{extractLT}
\alias{extractLTpwr}
\title{Obtain complete set of LT or LD estimates}
\usage{
extractLT(p = 0.99, obj, link = "logit", logscale = FALSE, nEsts = 8,
  slopeAdd = 8, eps = 0, scaling = 1, df.t = NULL)

extractLTpwr(p = 0.99, obj, link = "fpower", logscale = FALSE,
  nEsts = 8, slopeAdd = 8, lambda = 0, eps = 0.015, scaling = 1,
  df.t = NULL)
}
\arguments{
\item{p}{Desired mortality proportion.}

\item{obj}{\code{merMod} object, created using \code{lmer()} or
\code{glmer()}.}

\item{link}{Link function, specified as a character string.
For a folded power function, specify \code{"fpower"}.}

\item{logscale}{Logical.  Specify \code{TRUE}, if LT values are
to be back-transformed from a logarithmic scale.}

\item{nEsts}{Number of groups. The first \code{nEsts} estimates
are assumed to be intercepts.}

\item{slopeAdd}{Add to the intercept number to get corresponding slope.}

\item{eps}{Replace \code{prob} by \code{prob+eps} before transformation.}

\item{scaling}{Use to undo scaling of time or dose variable.}

\item{df.t}{Degrees of freedom for variance-covariance matrix.
If not supplied, it will be calculated internally.}

\item{lambda}{(\code{extractLTpwr} only) Power for power function.}
}
\value{
Matrix holding LD or LD estimates.
}
\description{
Obtain complete set of LT or LD estimates
}
