% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qra.r
\name{qra_weighted_average_interval_score}
\alias{qra_weighted_average_interval_score}
\title{Calculates an weighted interval score}
\usage{
qra_weighted_average_interval_score(
  weights,
  x,
  enforce_normalisation,
  per_centile_weights
)
}
\arguments{
\item{weights}{weights of the models}

\item{x}{input data frame}

\item{enforce_normalisation}{if TRUE, normalisation (weights sum to 1) is enforced}

\item{per_centile_weights}{if TRUE, separate weights are calculated for
different centiles}
}
\value{
average interval score
}
\description{
Calculates an weighted interval score
}
\keyword{internal}
